% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycodes.R
\name{countrycodes}
\alias{countrycodes}
\title{\emph{countrycodes}: retrieve country codes for the EatConscious package.}
\usage{
countrycodes(continent)
}
\arguments{
\item{continent}{String argument; for which continent would you like to retrieve the country codes?}
}
\value{
List of country codes within the specified continent.
}
\description{
\emph{countrycodes}: retrieve country codes for the EatConscious package.
}
