% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_functions.R
\name{plot_nutrients}
\alias{plot_nutrients}
\title{\emph{plot_nutrients}: plot the nutritional information of a specified food item.}
\usage{
plot_nutrients(food)
}
\arguments{
\item{food}{The food item as a string.}
}
\value{
Two pie charts with the percentage of nutrients based on the recommended daily intake that one serving of the food item gives.
}
\description{
\emph{plot_nutrients}: plot the nutritional information of a specified food item.
}
\examples{
plot_nutrients("potato")

}
