% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_food.R
\name{search_food}
\alias{search_food}
\title{\emph{search_food}: search for food items.}
\usage{
search_food(food, picture = TRUE)
}
\arguments{
\item{food}{The food item, input as a string, that you would like to search for.}

\item{picture}{Logical argument. Should a picture accompany the search result? Default = TRUE.}
}
\value{
A list of nutritional and climate-impact information for each search hit and a picture of the first hit.
}
\description{
\emph{search_food}: search for food items.
}
\examples{
search_food("potato")

}
