% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wf_data}
\alias{wf_data}
\title{Water Footprint Dataset from the Water Footprint Network.}
\format{
A data frame where each food item has a column for:
\describe{
  \item{The global average water footprint}{}
  \item{The average water footprint for each country}{}
  \item{The average water footprint for each province / state within each country}{}

}
}
\source{
\url{https://waterfootprint.org/en/resources/waterstat/product-water-footprint-statistics/}
}
\usage{
wf_data
}
\description{
A dataset containing the blue, green, and grey water footprint of various food items.
}
\keyword{datasets}
