% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_global_plot}
\alias{wf_global_plot}
\title{\emph{plot_wf_global}: visualize the global water footprint of a food item.}
\usage{
wf_global_plot(food)
}
\arguments{
\item{food}{Any food item. Input as string.}
}
\value{
A plot with the global blue, green, and grey water footprint of the food item.
}
\description{
\emph{plot_wf_global}: visualize the global water footprint of a food item.
}
\examples{
plot_wf_global(potato)

}
