% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfootprint_functions.R
\name{wf_region_plot}
\alias{wf_region_plot}
\title{\emph{wf_region_plot}: plot the water footprint of a food item across regions of a country.}
\usage{
wf_region_plot(food, country)
}
\arguments{
\item{food}{Any food item. Input as string.}

\item{country}{Input country as string. Must be entered as full name of country in Sentence Case.}
}
\value{
A plot of the blue, green, and grey water footprint of the food item per region in the specified country.
}
\description{
\emph{wf_region_plot}: plot the water footprint of a food item across regions of a country.
}
\examples{
wf_region_plot(potato, Netherlands)

}
